# docs/swagger.yaml
basePath: /
definitions:
  domain.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: string
    type: object
  handler.CreateCategoryInput:
    properties:
      color:
        type: string
      name:
        minLength: 3
        type: string
    required:
    - name
    type: object
  handler.CreateTodoInput:
    properties:
      description:
        minLength: 3
        type: string
      title:
        minLength: 3
        type: string
    required:
    - title
    type: object
  handler.LoginInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  handler.RegisterInput:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  handler.UpdateCategoryInput:
    properties:
      color:
        type: string
      name:
        minLength: 3
        type: string
    required:
    - name
    type: object
  handler.UpdatePasswordInput:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        minLength: 8
        type: string
    required:
    - new_password
    - old_password
    type: object
  handler.UpdateProfileInput:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
    required:
    - email
    - name
    type: object
  handler.UpdateTodoInput:
    properties:
      completed:
        type: boolean
      description:
        minLength: 3
        type: string
      title:
        minLength: 3
        type: string
    required:
    - title
    type: object
host: localhost:8080
info:
  contact:
    email: ewink@gmail.com
    name: Ewink
  description: REST API untuk manajemen Todo dengan autentikasi JWT
  title: Todo API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login dengan email dan password untuk mendapatkan JWT token
      parameters:
      - description: Data login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan user baru dengan nama, email, dan password
      parameters:
      - description: Data registrasi
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Register user baru
      tags:
      - auth
  /categories:
    get:
      description: Mengambil semua kategori milik user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Ambil semua kategori
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Membuat kategori baru untuk user yang sedang login
      parameters:
      - description: Data kategori
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Buat kategori baru
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Menghapus kategori milik user berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Hapus kategori
      tags:
      - categories
    get:
      description: Mengambil satu kategori milik user berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Ambil kategori berdasarkan ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Mengupdate kategori milik user berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data kategori
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Update kategori
      tags:
      - categories
  /todos:
    get:
      description: Mengambil semua todo milik user dengan pagination dan filter
      parameters:
      - description: 'Halaman ke berapa (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah data per halaman (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Cari berdasarkan judul
        in: query
        name: search
        type: string
      - description: Filter berdasarkan status
        in: query
        name: completed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Ambil semua todo
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Membuat todo baru untuk user yang sedang login
      parameters:
      - description: Data todo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTodoInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Buat todo baru
      tags:
      - todos
  /todos/{id}:
    delete:
      description: Menghapus todo milik user berdasarkan ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Hapus todo
      tags:
      - todos
    get:
      description: Mengambil satu todo milik user berdasarkan ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Ambil todo berdasarkan ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Mengupdate todo milik user berdasarkan ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data todo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTodoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Update todo
      tags:
      - todos
  /users/me:
    get:
      description: Mengambil data profil user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Ambil profil user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Mengupdate nama dan email user yang sedang login
      parameters:
      - description: Data profil
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Update profil user
      tags:
      - users
  /users/me/password:
    put:
      consumes:
      - application/json
      description: Mengganti password user yang sedang login
      parameters:
      - description: Data password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdatePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Ganti password
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"